Conflicts:
	BrainPortal/app/controllers/application_controller.rb
	BrainPortal/app/controllers/users_controller.rb
	BrainPortal/app/helpers/data_providers_helper.rb
	BrainPortal/app/models/site.rb
	BrainPortal/app/models/tool.rb
	BrainPortal/app/models/userfile_custom_filter.rb
	BrainPortal/lib/tasks/rspec.rake

+++ BrainPortal/app/models/userfile_custom_filter.rb
@@ -85,7 +85,7 @@ class UserfileCustomFilter < CustomFilter
   
   #Return +scope+ modified to filter the Userfile entry's size.
   def scope_size(scope)
-    scope.where( ["userfiles.size #{inequality_type(self.data["size_type"])} ?", (self.data["size_term"].to_i * 1000)])
+    scope.where( ["userfiles.size #{inequality_type(self.data["size_type"])} ?", (self.data["size_term"].to_f * 1000)])
   end
   
   #Return +scope+ modified to filter the Userfile entry's owner.
